+++ src/tests/search-tavily.test.ts
@@ -17,6 +17,7 @@ import assert from "node:assert/strict";
 import { resolveSearchProvider } from "../resources/extensions/search-the-web/provider.ts";
 import { normalizeQuery } from "../resources/extensions/search-the-web/url-utils.ts";
 import { mapFreshnessToTavily } from "../resources/extensions/search-the-web/tavily.ts";
+import { normalizeHeaders, parseJsonBody } from "./fetch-test-helpers.ts";
 
 // =============================================================================
 // Helpers for mocking global fetch
@@ -45,22 +46,6 @@ function makeTavilyResponse(overrides: Record<string, unknown> = {}) {
     response_time: "0.5",
     ...overrides,
   };
-}
-
-function normalizeHeaders(headers: HeadersInit | undefined): Record<string, string> | undefined {
-  if (headers == null) return undefined;
-  if (headers instanceof Headers) {
-    const result: Record<string, string> = {};
-    headers.forEach((v, k) => { result[k] = v; });
-    return result;
-  }
-  if (Array.isArray(headers)) return Object.fromEntries(headers);
-  return headers as Record<string, string>;
-}
-
-function parseJsonBody(body: BodyInit | null | undefined): Record<string, unknown> | undefined {
-  if (body == null || typeof body !== "string") return undefined;
-  try { return JSON.parse(body); } catch { return undefined; }
 }
 
 /**
